chore: refactor icons-view with vite and vue3
