--- packages/view/babel.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  presets: [['@vue/cli-plugin-babel/preset', { jsx: false }],
-    ['@vue/babel-preset-jsx', { compositionAPI: true }]],
-};
